requirements.txt
# Snowflake connectivity
snowflake-connector-python>=3.6.0
snowflake-snowpark-python>=1.11.0

# Streamlit for UI
streamlit>=1.31.0

# Snowflake Streamlit connector (simplifies auth)
snowflake-ml-python>=1.5.0

